+++ src/functionality.js
@@ -1,4 +1,5 @@
 /* eslint-disable no-use-before-define */
+import { updateTask, handleCheckbox, focus, blur } from "./updateTask";
 
 const removeTask = (task) => {
   const tasksListEl = task.parentNode;
@@ -12,29 +13,6 @@ const removeTask = (task) => {
   });
   localStorage.setItem('tasks', JSON.stringify(newTasksList));
   createTasksList();
-};
-
-const focus = (task) => {
-  task.classList.add('focus');
-  const deleteIcon = task.lastElementChild;
-  const dotsIcon = task.querySelector('.dots-icon');
-  dotsIcon.classList.remove('active');
-  deleteIcon.classList.add('active');
-};
-
-const blur = (task) => {
-  task.classList.remove('focus');
-  const deleteIcon = task.lastElementChild;
-  const dotsIcon = task.querySelector('.dots-icon');
-  dotsIcon.classList.add('active');
-  deleteIcon.classList.remove('active');
-};
-
-export const handleCheckbox = (checkbox) => {
-  const tasks = JSON.parse(localStorage.getItem('tasks'));
-  const index = parseInt(checkbox.id.slice(-1), 10);
-  tasks[index - 1].completed = checkbox.checked;
-  localStorage.setItem('tasks', JSON.stringify(tasks));
 };
 
 export const addTaskDom = (description, index, completed) => {
@@ -54,6 +32,9 @@ export const addTaskDom = (description, index, completed) => {
   });
   newTask.querySelector('input[type=text]').addEventListener('focus', (e) => {
     focus(e.target.parentNode);
+  });
+  newTask.querySelector('input[type=text]').addEventListener('change', (e) => {
+    updateTask(e.target);
   });
   newTask.querySelector('input[type=text]').addEventListener('blur', (e) => {
     blur(e.target.parentNode);
@@ -113,12 +94,4 @@ export const removeCompletedTasks = () => {
     });
     localStorage.setItem('tasks', JSON.stringify(newTasksList));
   }
-};
-
-export const updateTask = (input) => {
-  const tasksList = JSON.parse(localStorage.getItem('tasks'));
-  const taskId = input.parentNode.id;
-  const index = parseInt(taskId.slice(-1), 10);
-  tasksList[index - 1].description = input.value;
-  localStorage.setItem('tasks', JSON.stringify(tasksList));
 };
